// postgre-server/prisma/generated/zod/inputTypeSchemas/PinterestDataMinOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';

export const PinterestDataMinOrderByAggregateInputSchema: z.ZodType<Prisma.PinterestDataMinOrderByAggregateInput> = z.object({
  id: z.lazy(() => SortOrderSchema).optional(),
  username: z.lazy(() => SortOrderSchema).optional(),
  postId: z.lazy(() => SortOrderSchema).optional(),
  documentId: z.lazy(() => SortOrderSchema).optional()
}).strict();

export default PinterestDataMinOrderByAggregateInputSchema;